pyproject.toml
[project]
name = "fastapiagentparking"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "load-dotenv>=0.1.0",
    "pydantic[email]>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "router>=0.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
